+++ src/Bootstrap/DatabaseFoundation.php
@@ -9,6 +9,7 @@ use EventFlow\Application\Attendee\AttendeeService;
 use EventFlow\Application\Attendee\AttendeeQueryService;
 use EventFlow\Application\CheckIn\CheckInService;
 use EventFlow\Application\Communication\CommunicationService;
+use EventFlow\Application\Communication\TemplateAccessService;
 use EventFlow\Application\Communication\TemplateRenderer;
 use EventFlow\Application\Authorization\AuthorizationService;
 use EventFlow\Application\Authorization\RoleCapabilityPolicy;
@@ -113,6 +114,7 @@ final readonly class DatabaseFoundation
         public SeatingGroupMoveService $seatingGroupMoves,
         public CheckInService $checkIn,
         public CommunicationService $communications,
+        public TemplateAccessService $templateAccess,
         public ExportService $exports,
         public PrivacyService $privacy,
         public ProviderService $providers,
@@ -159,6 +161,8 @@ final readonly class DatabaseFoundation
         $importRepository = new WpdbImportRepository($database, $tableNames);
         $jobRepository = new WpdbJobRepository($database, $tableNames);
         $seatingRepository = new WpdbSeatingRepository($database, $tableNames);
+        $communicationRepository = new WpdbCommunicationRepository($database, $tableNames);
+        $templateRenderer = new TemplateRenderer();
         $exportStorage = new WordPressProtectedExportStorage($shared->random);
         $privacy = new PrivacyService(
             new WpdbPrivacyRepository($database, $tableNames),
@@ -336,12 +340,20 @@ final readonly class DatabaseFoundation
                 $shared->random,
             ),
             communications: new CommunicationService(
-                new WpdbCommunicationRepository($database, $tableNames),
+                $communicationRepository,
                 $authorization,
                 $idempotency,
                 $audit,
                 $shared->clock,
-                new TemplateRenderer(),
+                $templateRenderer,
+            ),
+            templateAccess: new TemplateAccessService(
+                $communicationRepository,
+                $authorization,
+                $idempotency,
+                $audit,
+                $shared->clock,
+                $templateRenderer,
             ),
             exports: new ExportService(
                 new WpdbExportRepository($database, $tableNames),